requirements.txt
asgiref==3.4.1
Django==3.2.6
django-smtp-ssl==1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
psycopg2==2.9.1
PyJWT==2.1.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
